refactor(audit): move package to internal/audit/
Move the audit subsystem from internal/access/policy/audit/ to
internal/audit/ to reflect its role as a general-purpose decision
recording facility rather than a policy-engine-owned subsystem.

Pure mechanical move: Entry type, field names, and all semantics
are unchanged in this commit. Follow-up commits rename fields and
add source-agnostic fields (Source, Component, Message).

Bead: holomush-ggbz
