+++ internal/access/policy/engine_test.go
@@ -19,8 +19,8 @@ import (
 
 	"github.com/holomush/holomush/internal/access"
 	"github.com/holomush/holomush/internal/access/policy/attribute"
-	"github.com/holomush/holomush/internal/access/policy/audit"
 	"github.com/holomush/holomush/internal/access/policy/types"
+	"github.com/holomush/holomush/internal/audit"
 )
 
 // mockSessionResolver is a test double for SessionResolver.
